icons in msgs are clickable now
