+++ client/src/Chat/ChatList.jsx
@@ -1,9 +1,11 @@
+/* eslint-disable react/no-array-index-key */
+/* eslint-disable react/prop-types */
 import React, {useRef, useState, useEffect} from 'react';
 import styled from 'styled-components';
 import Message from './Message';
 import ChatBar from './ChatBar';
 
-function ChatList({ chatRoomData, user, socket }) {
+function ChatList({ chatRoomData, user, socket, userForModal, setShowUserModal, getUserInfo }) {
   const chatRef = useRef();
   const [message, setMessage] = useState('');
 
@@ -14,7 +16,16 @@ function ChatList({ chatRoomData, user, socket }) {
   return (
     <ChatRoomMain>
       <ChatContainer ref={chatRef}>
-        {chatRoomData.length > 0 && chatRoomData.map((message, index) => <Message key={`${message?.email}-${message?.timestamp}-${index}`} user={user} message={message} />)}
+        {chatRoomData.length > 0
+          && chatRoomData.map((message, index) =>
+            <Message
+              key={`${message?.email}-${message?.timestamp}-${index}`}
+              user={user}
+              message={message}
+              getUserInfo={getUserInfo}
+              userForModal={userForModal}
+              setShowUserModal={setShowUserModal}
+            />)}
       </ChatContainer>
       <ChatBar user={user} socket={socket} />
     </ChatRoomMain>
