+++ client/src/Games/ScoreBoard.jsx
@@ -2,16 +2,20 @@
 /* eslint-disable react/prop-types */
 import React from 'react';
 import styled from 'styled-components';
-import { Button } from '@mui/material';
-import getGmailUsername from './lib/getGmailUsername';
+import getGmailUsername from '../lib/getGmailUsername';
 
-function FinalScore({ scores, setScoreShowing }) {
-  const back = () => {
-    setScoreShowing(false);
-  };
+function ScoreBoard({ scores }) {
   return (
     <ScoreContainer>
-      <Title>final scores.</Title>
+      {/* {scores
+        .map((score, index) => (
+          <Score key={`${score.displayName}-${index}`}>
+            <ScoreInner>
+              <Name>{`${score.displayName}: `}</Name>
+              <Count>{`${score.count}`}</Count>
+            </ScoreInner>
+          </Score>
+        ))} */}
       {scores.map((score, index) => (
         <Score key={`${score.displayName}-${index}`}>
           <Left>
@@ -28,12 +32,11 @@ function FinalScore({ scores, setScoreShowing }) {
           </Div>
         </Score>
       ))}
-      <Button onClick={back}>back</Button>
     </ScoreContainer>
   );
 }
 
-export default FinalScore;
+export default ScoreBoard;
 
 const Div = styled.div`
   display: flex;
@@ -64,12 +67,25 @@ const Img = styled.img`
   margin-left: 30px;
 `;
 
-const Button1 = styled.button`
-  width: 200px;
-  height: 100px;
-  padding: 3px;
-  font-size: 50px;
-  text-align: center;
+
+const Name = styled.p`
+  font-weight: normal;
+  margin-left: 2%;
+  overflow: hidden;
+`;
+
+const Count = styled.p`
+  margin-right: 2%;
+  font-weight: bold;
+  overflow: hidden;
+`;
+
+const ScoreInner = styled.div`
+  width: 50%;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
 `;
 
 const Score = styled.div`
@@ -79,57 +95,18 @@ const Score = styled.div`
   flex-direction: row;
   justify-content: space-between;
   align-items: center;
-  z-index:101;
   border: .3px solid black;
   margin-bottom: 2px;
 `;
-
-const Title = styled.p`
-  font-weight: bold;
-  letter-spacing: -1px;
-  font-size: 40px;
-`;
-const Name = styled.p`
-  font-weight: normal;
-  margin-left: 2%;
-`;
-
-const Count = styled.p`
-  margin-right: 2%;
-  font-weight: bold;
-`;
-
-// const Score = styled.div`
-//   display: flex;
-//   padding: 3%;
-//   width: 90%;
-//   font-size: 20px;
-//   /* margin-left: 1%; */
-//   display: flex;
-//   flex-direction: row;
-//   justify-content: space-between;
-//   align-items: baseline;
-//   border: .3px solid black;
-//   margin-bottom: 1%;
-// `;
 
 const ScoreContainer = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
   margin-top:1%;
-  width: 96%;
+  width: 100%;
   overflow-y: auto;
   align-items: center;
-  cursor: default;
-
-  > button {
-    margin-top: 3%;
-    background-color: gainsboro !important;
-    color: black;
-    border: none;
-  }
-  > button:hover {
-    opacity: 70%;
-  }
+  /* border-top: .3px solid black; */
+  overflow: hidden;
 `;
