+++ client/src/Menu/StatsTile.jsx
@@ -1,15 +1,16 @@
-import React, {useState} from 'react';
+/* eslint-disable react/prop-types */
+import React, { useState } from 'react';
 import styled from 'styled-components';
 import { format, parseISO } from 'date-fns';
-import getGmailUsername from './lib/getGmailUsername';
-import getWinPerc from './lib/getWinPerc';
-import UserInfo from './UserInfo';
+import getGmailUsername from '../lib/getGmailUsername';
+import getWinPerc from '../lib/getWinPerc';
+import UserInfo from '../Users/UserInfo';
 
 function StatsTile({ user }) {
   const [showModal, setShowModal] = useState(false);
   const handleClick = () => {
     setShowModal(true);
-  }
+  };
   return (
     <>
       <User onClick={handleClick}>
@@ -28,7 +29,7 @@ function StatsTile({ user }) {
 export default StatsTile;
 
 const User = styled.div`
-  width: 98%;
+  width: 97%;
   font-size: 15px;
   display: flex;
   flex-direction: row;
@@ -36,6 +37,7 @@ const User = styled.div`
   justify-content: space-between;
   &:hover {
     opacity: 90%;
+    border: .1px solid black;
     cursor: pointer;
     background: #efefef;
     transition: 0.3s;
@@ -45,15 +47,15 @@ const User = styled.div`
 `;
 
 const Name = styled.div`
-  width: 20%;
+  width: 32%;
   text-align: center;
-`
+`;
 
 const Numbers = styled.div`
   width: 10%;
   text-align: center;
-`
+`;
 const Right = styled.div`
-  width: 15%;
+  width: 18%;
   text-align: center;
-`
+`;
