+++ package.json
@@ -3,7 +3,6 @@
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
-
   "scripts": {
     "start": "webpack --mode=development --watch",
     "server-dev": "npx nodemon server/index.js",
@@ -37,7 +36,6 @@
     "socket.io": "^4.5.1",
     "socket.io-client": "^4.5.1",
     "styled-components": "^5.3.5",
-    "underscore": "^1.13.4",
     "use-timer": "^2.0.1"
   },
   "devDependencies": {
